% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_training.R
\name{append_results}
\alias{append_results}
\title{Make or append a result list with  placeholders if needed.}
\usage{
append_results(inc_methods, init_list = NULL)
}
\description{
Make or append a result list with  placeholders if needed.
}
