% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_training.R
\name{autofit}
\alias{autofit}
\title{Automatically trains multiple models for a fixed time}
\usage{
autofit(
  traindata,
  target_name,
  result_list,
  max_time = 5 * 60,
  n_kernel = 1,
  ...
)
}
\description{
Automatically trains multiple models for a fixed time
}
