\name{CalcConPwrExpAndersen}

\alias{CalcConPwrExpAndersen}

\title{
 Auxiliary Function
}

\description{
 Calculates the conditional power in the exponential model
 with the original formulae of the Andersen paper.
}

\author{
 Andreas Kuehnapfel
}

\seealso{
 \code{\link{CP}}
 \code{\link{ConPwrExp}}
 \code{\link{CompSurvMod}}
}
