\name{CalcConPwrNonMix}

\alias{CalcConPwrNonMix}

\title{
 Auxiliary Function
}

\description{
 Calculates the conditional power in the non-mixture models.
}

\author{
 Andreas Kuehnapfel
}

\seealso{
 \code{\link{CP}}
 \code{\link{ConPwrNonMixExp}}
 \code{\link{ConPwrNonMixWei}}
 \code{\link{ConPwrNonMixGamma}}
 \code{\link{CompSurvMod}}
}
