\name{DispConPwr}

\alias{DispConPwr}

\title{
 Auxiliary Function
}

\description{
 Displays the conditional power in
 \code{\link{ConPwrExp}}, \code{\link{ConPwrNonMixExp}},
 \code{\link{ConPwrNonMixWei}} and \code{\link{ConPwrNonMixGamma}}.
}

\author{
 Andreas Kuehnapfel
}

\seealso{
 \code{\link{CP}}
 \code{\link{ConPwrExp}}
 \code{\link{ConPwrNonMixExp}}
 \code{\link{ConPwrNonMixWei}}
 \code{\link{ConPwrNonMixGamma}}
}
