\name{FctPersMonNonMixExp}

\alias{FctPersMonNonMixExp}

\title{
 Auxiliary Function
}

\description{
 Calculates the value of some function of the further person months
 in the non-mixture model with exponential survival.
}

\author{
 Andreas Kuehnapfel
}

\seealso{
 \code{\link{CP}}
 \code{\link{ConPwrNonMixExp}}
}
