\name{FctPersMonNonMixGamma}

\alias{FctPersMonNonMixGamma}

\title{
 Auxiliary Function
}

\description{
 Calculates the value of some function of the further person months
 in the non-mixture model with Gamma type survival.
}

\author{
 Andreas Kuehnapfel
}

\seealso{
 \code{\link{CP}}
 \code{\link{ConPwrNonMixGamma}}
}
