\name{FctPersMonNonMixWei}

\alias{FctPersMonNonMixWei}

\title{
 Auxiliary Function
}

\description{
 Calculates the value of some function of the further person months
 in the non-mixture model with Weibull type survival.
}

\author{
 Andreas Kuehnapfel
}

\seealso{
 \code{\link{CP}}
 \code{\link{ConPwrNonMixWei}}
}
