\name{InitValLikelihoodNonMixExp}

\alias{InitValLikelihoodNonMixExp}

\title{
 Auxiliary Function
}

\description{
 Calculates initial values for maximum likelihood calculations
 in the non-mixture model with exponential survival.
}

\author{
 Andreas Kuehnapfel
}

\seealso{
 \code{\link{CP}}
 \code{\link{ConPwrNonMixExp}}
}
