\name{InitValLikelihoodNonMixGamma}

\alias{InitValLikelihoodNonMixGamma}

\title{
 Auxiliary Function
}

\description{
 Calculates initial values for maximum likelihood calculations
 in the non-mixture model with Gamma type survival.
}

\author{
 Andreas Kuehnapfel
}

\seealso{
 \code{\link{CP}}
 \code{\link{ConPwrNonMixGamma}}
}
