\name{InterimData}

\alias{InterimData}

\title{
 Auxiliary Function
}

\description{
 Displays the data from the interim analysis.
}

\author{
 Andreas Kuehnapfel
}

\seealso{
 \code{\link{CP}}
 \code{\link{ConPwrExp}}
 \code{\link{ConPwrNonMixExp}}
 \code{\link{ConPwrNonMixWei}}
 \code{\link{ConPwrNonMixGamma}}
 \code{\link{CompSurvMod}}
}
