\name{IsValid}

\alias{IsValid}

\title{
 Auxiliary Function
}

\description{
 Checks the passed parameters of the user.
}

\author{
 Andreas Kuehnapfel
}

\seealso{
 \code{\link{CP}}
 \code{\link{ConPwrExp}}
 \code{\link{ConPwrNonMixExp}}
 \code{\link{ConPwrNonMixWei}}
 \code{\link{ConPwrNonMixGamma}}
 \code{\link{CompSurvMod}}
}
