\name{LikelihoodNonMixExp}

\alias{LikelihoodNonMixExp}

\title{
 Auxiliary Function
}

\description{
 Calculates the maximum likelihood estimators
 of the non-mixture model with exponential survival.
}

\author{
 Andreas Kuehnapfel
}

\seealso{
 \code{\link{CP}}
 \code{\link{ConPwrNonMixExp}}
}
