\name{LikelihoodNonMixGamma}

\alias{LikelihoodNonMixGamma}

\title{
 Auxiliary Function
}

\description{
 Calculates the maximum likelihood estimators
 of the non-mixture model with Gamma type survival.
}

\author{
 Andreas Kuehnapfel
}

\seealso{
 \code{\link{CP}}
 \code{\link{ConPwrNonMixGamma}}
}
