\name{LikelihoodNonMixWei}

\alias{LikelihoodNonMixWei}

\title{
 Auxiliary Function
}

\description{
 Calculates the maximum likelihood estimators
 of the non-mixture model with Weibull type survival.
}

\author{
 Andreas Kuehnapfel
}

\seealso{
 \code{\link{CP}}
 \code{\link{ConPwrNonMixWei}}
}
