\name{PersMonNonMixExp}

\alias{PersMonNonMixExp}

\title{
 Auxiliary Function
}

\description{
 Calculates the further person months
 in the non-mixture model with exponential survival.
}

\author{
 Andreas Kuehnapfel
}

\seealso{
 \code{\link{CP}}
 \code{\link{ConPwrNonMixExp}}
}
