\name{PersMonNonMixGamma}

\alias{PersMonNonMixGamma}

\title{
 Auxiliary Function
}

\description{
 Calculates the further person months
 in the non-mixture model with Gamma type survival.
}

\author{
 Andreas Kuehnapfel
}

\seealso{
 \code{\link{CP}}
 \code{\link{ConPwrNonMixGamma}}
}
