\name{PersMonNonMixWei}

\alias{PersMonNonMixWei}

\title{
 Auxiliary Function
}

\description{
 Calculates the further person months
 in the non-mixture model with Weibull type survival.
}

\author{
 Andreas Kuehnapfel
}

\seealso{
 \code{\link{CP}}
 \code{\link{ConPwrNonMixWei}}
}
