\name{PlotConPwr}

\alias{PlotConPwr}

\title{
 Auxiliary Function
}

\description{
 Plots the conditional power curve in
 \code{\link{ConPwrExp}}, \code{\link{ConPwrNonMixExp}},
 \code{\link{ConPwrNonMixWei}} and \code{\link{ConPwrNonMixGamma}}.
}

\author{
 Andreas Kuehnapfel
}

\seealso{
 \code{\link{CP}}
 \code{\link{ConPwrExp}}
 \code{\link{ConPwrNonMixExp}}
 \code{\link{ConPwrNonMixWei}}
 \code{\link{ConPwrNonMixGamma}}
}
