\name{PlotEstNonMixExp}

\alias{PlotEstNonMixExp}

\title{
 Auxiliary Function
}

\description{
 Plots the estimated survival curves
 of the non-mixture model with exponential survival.
}

\author{
 Andreas Kuehnapfel
}

\seealso{
 \code{\link{CP}}
 \code{\link{ConPwrNonMixExp}}
}
