\name{PlotEstNonMixGamma}

\alias{PlotEstNonMixGamma}

\title{
 Auxiliary Function
}

\description{
 Plots the estimated survival curves
 of the non-mixture model with Gamma type survival.
}

\author{
 Andreas Kuehnapfel
}

\seealso{
 \code{\link{CP}}
 \code{\link{ConPwrNonMixGamma}}
}
