\name{PlotEstNonMixWei}

\alias{PlotEstNonMixWei}

\title{
 Auxiliary Function
}

\description{
 Plots the estimated survival curves
 of the non-mixture model with Weibull type survival.
}

\author{
 Andreas Kuehnapfel
}

\seealso{
 \code{\link{CP}}
 \code{\link{ConPwrNonMixWei}}
}
