\name{PlotKM}

\alias{PlotKM}

\title{
 Auxiliary Function
}

\description{
 Plots the Kaplan-Meier curves.
}

\author{
 Andreas Kuehnapfel
}

\seealso{
 \code{\link{CP}}
 \code{\link{ConPwrExp}}
 \code{\link{ConPwrNonMixExp}}
 \code{\link{ConPwrNonMixWei}}
 \code{\link{ConPwrNonMixGamma}}
 \code{\link{CompSurvMod}}
}
