% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess_necessary}
\alias{preprocess_necessary}
\title{Necessary Preprocessing}
\usage{
preprocess_necessary(dataset, missings_threshold = 0.5, const_imputation = -1)
}
\arguments{
\item{dataset}{A dataframe.}

\item{missings_threshold}{Missings threshold in [0,1] (if NULL, no variable will be removed).}

\item{const_imputation}{Value which will replace each missing (if NULL, no value will be imputed).}
}
\value{
A cleaned dataset.
}
\description{
Removes constant variables and optionally variables with too many missing and imputes missing values by a constant.
}
