\name{SplitData}

\alias{SplitData}

\title{
 Auxiliary Function
}

\description{
 Splits the entire data frame into two sub data frames
 each for one group.
}

\author{
 Andreas Kuehnapfel
}

\seealso{
 \code{\link{CP}}
 \code{\link{ConPwrExp}}
 \code{\link{ConPwrNonMixExp}}
 \code{\link{ConPwrNonMixWei}}
 \code{\link{ConPwrNonMixGamma}}
 \code{\link{CompSurvMod}}
}
