\name{test}

\docType{data}

\alias{test}

\title{
 Test Data Frame
}

\description{
 The data frame 'test' is generated by random
 and does not refer to a special realistic issue.
}

\format{
 This data frame consists of three columns.
 The first column consists of the group expressions 'A' and 'B' (character).
 The second column consists of
 the status 1 for event or 2 for censored (numeric).
 The third column consists of the event time (numeric). 
}

\source{
 \code{\link{CP}} \cr
 \code{\link{GenerateDataFrame}}
}

\keyword{
 datasets
}