% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_training.R
\name{train}
\alias{train}
\title{Custom wrapper for caret::train}
\usage{
train(data, target, max_time = 2 * 60 * 60, n_kernel = 1, ...)
}
\description{
includes choosable timeout, target is passed as name, other standard trControl
TODO: multiple targets
}
